Collision Detection sorta

+++ src/entities.js
@@ -4,20 +4,33 @@ function Entities({scene, onAdd, onRemove}) {
   return {
     e,
     add: (o) => {
-      // o.tag === "bullet" && console.log("EntityAdd", o.tag, o);
       if (!Array.isArray(o.tags) && o.tag) {
         o.tags = [o.tag];
+        delete o.tag;
       }
+      if (!o.id) o.id = o.mesh?.uuid || 0;
       if (o.id) indexById[o.id] = o;
       e.push(o);
       if (o.mesh) scene.add(o.mesh);
       if (typeof onAdd === 'function') onAdd(o);
     },
-    remove: (o) => {
-      if (o.id) delete indexById[o.id];
+    removeById: (id) => {
+      let o = indexById[id];
+      console.log('removeById', id, o);
       const index = e.indexOf(o);
+      if (!index) throw new Error('Entities cannot remove unknown object', o);
       if (index) e.splice(index, 1);
       if (typeof onRemove === 'function') onRemove(o);
+      if (o.mesh) scene.remove(o.mesh);
+    },
+    remove: (o) => {
+      console.log('remove', o.tags, o);
+      delete indexById[o.id];
+      const index = e.indexOf(o);
+      if (!index) throw new Error('Entities cannot remove unknown object', o);
+      if (index) e.splice(index, 1);
+      //if (typeof onRemove === 'function') onRemove(o);
+      //if (o.mesh) scene.remove(o.mesh);
     },
     getById: (id) => {
       return indexById[id];
